+++ adapter/taurusx/CHANGELOG.md
@@ -8,3 +8,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.12.2.0]
 - Updated TaurusX SDK to 1.12.2
+- [BDN-1093](https://appodeal.atlassian.net/browse/BDN-1093) Fixed revenue calculation for TaurusX adapter